requirements.txt
selenium>=4.23
python-dotenv>=1.0
webdriver-manager>=4.0
pyyaml>=6.0
pandas>=2.2
